Setup directory and templates for a new research study

This function can be used to automatically setup your study directory by creating folders and template scripts

create_study(type = "data analysis", sessions = 1,
  figures.dir = FALSE, manuscript.dir = FALSE,
  presentations.dir = FALSE, other.dir = c(), masterscript = NULL,
  rawscript = FALSE, scorescript = FALSE, mergescript = FALSE,
  path = ".")

Arguments

type
String. Options: "data collection"; "analysis".

sessions
Numeric. How many sessions will the study have?

figures.dir
Logical. Create figures directory? default = FALSE

manuscript.dir
Logical. Create manuscript directory? default = FALSE

presentations.dir
Logical. Create presentations directory? default = FALSE

other.dir
List of other directories you want to create

masterscript
Logical. Download masterscript template? default = TRUE

rawscript
Logical. Download rawscript template? default = FALSE

scorescript
Logical. Download scorescript template? default = FALSE

mergescript
Logical. Download mergescript template? default = FALSE

path
String. Home directory file path

scripts.dir
Logical. Create script directory? default = TRUE

data.dir
Logical. Create data files directory? default = TRUE

raw.dir
Logical. Create raw? default = TRUE

messy.dir
Logical. Create raw messy directory? default = TRUE

messy.name
String. Name of messy direcotry. default = "messy"

scored.dir
Logical. Create scored directory? default = TRUE

tasks.dir
Logical. Create tasks directory? default = TRUE

documents.dir
Logical. Create documents directory? default = TRUE

results.dir
Logical. Create results directory? default = TRUE

generic
Logical. Download all generic templates? default = FALSE

demographics
Logical. Download demographics template? default = FALSE

gf
Logical. Do you want to download gf template files?

wmc
Logical. Do you want to download wmc template files?

ac
Logical. Do you want to download all attention control template files?

antisaccade
Logical. Do you want to download antisaccade template files?

stroop
Logical. Do you want to download stroop template files?

flanker
Logical. Do you want to download flanker template files?

stroopDL
Logical. Do you want to download stroopDL template files?

flankerDL
Logical. Do you want to download flankerDL template files?

va4
Logical. Do you want to download va4 template files?

sact
Logical. Do you want to download sact template files?